+++ tests/testthat/test-plot_source_wads.R
@@ -12,13 +12,17 @@ test_that("gives error if not qsip object", {
 test_that("gives error if group is not a column name in source_data", {
   expect_no_error(plot_source_wads(example_qsip_object, group = "Moisture"))
   expect_error(plot_source_wads(example_qsip_object, group = "not_a_column"),
-            "group must be a column name in source_data")
+            "not found in source metadata")
+})
+
+test_that("group works with multiple columns", {
+  expect_true(inherits(plot_source_wads(example_qsip_object, group = c("Moisture", "isotopolog")), "ggplot"))
 })
 
 test_that("titling works", {
   expect_no_error(plot_source_wads(example_qsip_object, title = "A title"))
   expect_error(plot_source_wads(example_qsip_object, group = "not_a_column"),
-            "group must be a column name in source_data")
+            "not found in source metadata")
 
   p <- plot_source_wads(example_qsip_object)
   expect_null(p$labels$title)